Source.cpp
Test Driver for Graph class.
Takes a full pathway to a txt file. The file must be formatted correctly. For this version, the file's format must be: the first line is the number of vertices in the graph. The following lines must be pairs of vertices (labeled as numbers) the first number being the starting vertex, followed by a space, then the destination vertex.